backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.6.0
